Add files via upload
Fixed the split and added the merge portion we need to find a way to compare the two types and how we will sort them. I am making a api.txt which should be easy to follow. 
Still Need:
get the values into data
read from the file
compare the values
output the data